basic4validation.py now uses functions from basic.py
